+++ spec/capybara/httpclient_json/driver_spec.rb
@@ -7,6 +7,7 @@ describe klass do
 
   it_should_behave_like 'driver'
   it_should_behave_like 'driver with header support'
+  it_should_behave_like 'driver with custom header support'
   it_should_behave_like 'driver with status code support'
   it_should_behave_like 'driver with cookies support'
   it_should_behave_like 'driver with infinite redirect detection'